+++ server/notebooks/views.py
@@ -34,15 +34,15 @@ def _get_iframe_src():
 def notebook_view(request, pk):
     notebook = get_object_or_404(Notebook, pk=pk)
     if 'revision' in request.GET:
-        notebook_content = get_object_or_404(NotebookRevision, pk=int(request.GET['revision']))
+        revision = get_object_or_404(NotebookRevision, pk=int(request.GET['revision']))
     else:
-        notebook_content = notebook.revisions.last()
+        revision = notebook.revisions.first()
     notebook_info = {
         'user_can_save': notebook.owner_id == request.user.id,
         'notebook_id': notebook.id,
-        'revision_id': notebook_content.id,
+        'revision_id': revision.id,
         'connectionMode': 'SERVER',
-        'title': notebook_content.title
+        'title': revision.title
     }
     if notebook.forked_from is not None:
         notebook_info['forked_from'] = notebook.forked_from.id
@@ -51,7 +51,7 @@ def notebook_view(request, pk):
     return render(request, 'notebook.html', {
         'user_info': _get_user_info_json(request.user),
         'notebook_info': notebook_info,
-        'jsmd': notebook_content.content,
+        'jsmd': revision.content,
         'iframe_src': _get_iframe_src()
     })
 
@@ -81,12 +81,12 @@ def notebook_revisions(request, pk):
          'size': len(file.content)}
         for file in File.objects.filter(notebook_id=pk).order_by('-last_updated')
     ]
-    revisions = list(reversed([{
+    revisions = list([{
         'id': revision.id,
         'notebookId': revision.notebook_id,
         'title': revision.title,
         'date': revision.created.isoformat(sep=' ')}
-        for revision in NotebookRevision.objects.filter(notebook_id=pk)]))
+        for revision in NotebookRevision.objects.filter(notebook_id=pk)])
     return render(request, '../templates/index.html', {
             'page_data': {
                 'userInfo': get_user_info_dict(request.user),