+++ src/eval-frame/actions/language-actions.js
@@ -1,10 +1,10 @@
 import { postMessageToEditor } from "../port-to-editor";
 import {
   appendToEvalHistory,
-  historyIdGen,
   updateValueInHistory,
   sendStatusResponseToEditor
 } from "./actions";
+import generateRandomId from "../../tools/generate-random-id";
 
 export function addLanguage(languageDefinition) {
   return {
@@ -84,7 +84,7 @@ function loadLanguagePlugin(pluginData, historyId, dispatch) {
 
 export function evaluateLanguagePlugin(pluginText, evalId) {
   return dispatch => {
-    const historyId = historyIdGen.nextId();
+    const historyId = generateRandomId();
     dispatch(
       appendToEvalHistory(null, pluginText, undefined, {
         historyId,
@@ -124,7 +124,7 @@ export function ensureLanguageAvailable(languageId, state, dispatch) {
   if (
     Object.prototype.hasOwnProperty.call(state.languageDefinitions, languageId)
   ) {
-    const historyId = historyIdGen.nextId();
+    const historyId = generateRandomId();
     dispatch(
       appendToEvalHistory(
         null,