+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - change nomenclature around user output reps _including in API_ (**breaking change**)
 - removed errant "0" in landing page
 - plugins now throw network errors to the iodide console and catch syntax errors within plugin loaders
+- implements a language selector menu on the right side of the console input
 
 # 0.0.3 (2019-02-04)